refactor: extract durable benchmark catalog
