+++ devtools/benchmark_catalog.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass
 
 from polylogue.scenarios import ScenarioMetadata
 
-from .benchmark_campaign import CAMPAIGNS as BENCHMARK_CAMPAIGNS
+from .benchmark_scenario_catalog import BENCHMARK_CAMPAIGNS
 from .synthetic_benchmark_catalog import SYNTHETIC_BENCHMARK_SCENARIOS
